# distsql: wrap RowFetcher.NextRow() in RowSource.Next()