This is preparatory work for stat collection. A stat object will be
inserted between a processor and its upstream RowSource data source.
To be able to collect stats this way in the tableReader,
RowFetcher.NextRow has to be wrapped in a RowSource.Next implementation.

Release note: None

Here are the benchmark differences, I'm unsure why this change offers a performance improvement:

```
BENCHES=BenchmarkTableReader PKG=./pkg/sql/distsqlrun ./scripts/bench master pstats | tee trresults
Comparing pstats (new) with master (old)

Writing to /var/folders/ry/6tnq_y2x6mn_pk6sy57jqpjw0000gn/T/tmp.qzgNP6Qe
Switching to master
+ make bench PKG=./pkg/sql/distsqlrun BENCHTIMEOUT=5m BENCHES=BenchmarkTableReader 'TESTFLAGS=-count 10 -benchmem'
Switching to pstats
+ make bench PKG=./pkg/sql/distsqlrun BENCHTIMEOUT=5m BENCHES=BenchmarkTableReader 'TESTFLAGS=-count 10 -benchmem'
name           old time/op    new time/op    delta
TableReader-8    5.58ms ± 6%    5.30ms ± 4%  -5.07%  (p=0.004 n=10+10)

name           old alloc/op   new alloc/op   delta
TableReader-8    1.95MB ± 0%    1.95MB ± 0%    ~     (p=0.315 n=10+10)

name           old allocs/op  new allocs/op  delta
TableReader-8       826 ±22%       824 ±15%    ~     (p=0.491 n=10+10)
```

The abstraction isn't as clean as I would like it to be because of the initialization necessary for the `RowFetcher` (`init` + `StartScan` in `newTableReader` and `Next()`). I would appreciate any ideas for alternative approaches.

@jordanlewis making you the reviewer until we figure out who will be reviewing my stats-related changes but feel free to re-assign.