@jordanlewis 2018-04-02
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Reviewed 1 of 1 files at r1.
Review status: all files reviewed at latest revision, all discussions resolved.

---

*[pkg/sql/distsqlrun/tablereader.go, line 221 at r1](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24412#-L95vlKnCcYHX4EpI1Wh:-L95vlKnCcYHX4EpI1Wi:b-x0nu9e) ([raw file](https://github.com/cockroachdb/cockroach/blob/df5c653ed1091e74d166e673bb5dccd37559c8ae/pkg/sql/distsqlrun/tablereader.go#L221)):*
> ```Go
> 		}
> 
> 		tr.input = rowFetcherWrapper{ctx: tr.ctx, RowFetcher: &tr.fetcher}
> ```

Can this go into `initRowFetcher` instead of this only-on-first-Next area?

---


*Comments from [Reviewable](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24412#-:-L95w97-7-kQqINQeSy3:bnfp4nl)*
<!-- Sent from Reviewable.io -->