@jordanlewis 2018-04-02
Review status: all files reviewed at latest revision, 1 unresolved discussion, some commit checks pending.

---

*[pkg/sql/distsqlrun/tablereader.go, line 221 at r1](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24412#-L95vlKnCcYHX4EpI1Wh:-L95zy81-p9eoMmmh8Zp:b3yy7zw) ([raw file](https://github.com/cockroachdb/cockroach/blob/df5c653ed1091e74d166e673bb5dccd37559c8ae/pkg/sql/distsqlrun/tablereader.go#L221)):*
<details><summary><i>Previously, asubiotto (Alfonso Subiotto Marqués) wrote…</i></summary><blockquote>

Unfortunately not because the processor's context is initialized in `maybeStart`.
</blockquote></details>

Ah, gotcha. Ok that's fine!

---


*Comments from [Reviewable](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24412)*
<!-- Sent from Reviewable.io -->